implementation StudentService avec create and get methods

+++ src/main/java/ca/cal/leandrose/repository/StudentRepository.java
@@ -3,4 +3,7 @@ package ca.cal.leandrose.repository;
 import ca.cal.leandrose.model.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {}
+
+public interface StudentRepository extends JpaRepository<Student, Long> {
+
+}
